change welcome screen to home screen add ViewScreen set up linking closes (Replace Welcome Screen with MarqueeInputScreen #14)
