+++ app/components/index.ts
@@ -10,3 +10,4 @@ export * from "./Text"
 export * from "./TextField"
 export * from "./Toggle"
 export * from "./EmptyState"
+export * from "./MarqueeText"
