+++ app/i18n/ko.ts
@@ -7,7 +7,7 @@ const ko: Translations = {
     back: "뒤로",
     logOut: "로그아웃",
   },
-  welcomeScreen: {
+  homeScreen: {
     postscript:
       "잠깐! — 지금 보시는 것은 아마도 당신의 앱의 모양새가 아닐겁니다. (디자이너분이 이렇게 건내주셨다면 모를까요. 만약에 그렇다면, 이대로 가져갑시다!) ",
     createMarquee: "나만의 마키를 만들어보세요!",
