+++ app/screens/index.ts
@@ -1,4 +1,4 @@
-export * from "./WelcomeScreen"
+export * from "./HomeScreen"
 
 export * from "./DemoCommunityScreen"
 export * from "./DemoDebugScreen"
@@ -6,4 +6,4 @@ export * from "./DemoShowroomScreen/DemoShowroomScreen"
 
 export * from "./ErrorScreen/ErrorBoundary"
 // export other screens here
-export * from "./MarqueeInputScreen"
+export * from "./ViewScreen"
